backend/test/requirements.txt
groq
chromadb
opencv-python-headless==4.11.0.86
numpy==1.22.4
Pillow
python-dotenv
facenet-pytorch
resemblyzer
torch==2.2.2
torchvision==0.17.2
ddgs
librosa
praat-parselmouth
soundfile
openai
py-feat>=0.6.0
matplotlib>=3.7.0
h5py>=3.6.0
numexpr>=2.8.4
scipy==1.11.4
